move read methods into read.go, write methods into write.go
